Добавил вторую версию бота
