+++ .gitignore
@@ -1,3 +1,4 @@
+settings.py
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
