+++ 8_ephem_bot2.py
@@ -1,0 +1,42 @@
+import ephem
+import logging
+from telegram.ext import Updater, CommandHandler, MessageHandler, Filters
+import settings
+
+logging.basicConfig(filename='bot.log', level=logging.INFO)
+
+PROXY = {'proxy_url': settings.PROXY_URL,
+    'urllib3_proxy_kwargs': {'username': settings.PROXY_USERNAME, 'password': settings.PROXY_PASSWORD}}
+
+def greet_user(update, context):
+    print("Вызван /start")
+    update.message.reply_text('Привет, пользователь!')
+
+def planet(update, context):
+    print("Вызван /planet Mars")     
+    update.message.reply_text("Планета Марс")
+   
+    mars = ephem.Mars('2022/02/20')
+    const = ephem.constellation(mars)
+    update.message.reply_text(const)
+
+
+def talk_to_me(update, context):
+    text = update.message.text
+    print(text)
+    update.message.reply_text(text)
+
+def main():
+    mybot = Updater(settings.API_KEY, use_context=True, request_kwargs=PROXY)
+
+    dp = mybot.dispatcher
+    dp.add_handler(CommandHandler("start", greet_user))
+    dp.add_handler(CommandHandler("planet_Mars", planet))
+    dp.add_handler(MessageHandler(Filters.text, talk_to_me))
+
+    logging.info('Бот стартовал')
+    mybot.start_polling()
+    mybot.idle()
+
+if __name__=='__main__':    
+    main()
